refactor (service and controller for sockets into the correct folder)
