+++ src/main/java/ch/uzh/ifi/hase/soprafs23/service/SocketService.java
@@ -1,4 +1,4 @@
-package ch.uzh.ifi.hase.soprafs23;
+package ch.uzh.ifi.hase.soprafs23.service;
 
 
 import ch.uzh.ifi.hase.soprafs23.constant.MessageType;
